+++ src/Stormancer.Plugins/Steam/cpp/Steam.hpp
@@ -216,20 +216,23 @@ namespace Stormancer
 			{
 			public:
 
-				SteamState(std::shared_ptr<Configuration> config)
+				SteamState(std::shared_ptr<Configuration> config, std::shared_ptr<ILogger> logger)
 				{
 					_authenticationEnabled = config->additionalParameters.find(ConfigurationKeys::AuthenticationEnabled) != config->additionalParameters.end() ? (config->additionalParameters.at(ConfigurationKeys::AuthenticationEnabled) != "false") : true;
 					_connectLobby = config->additionalParameters.find(ConfigurationKeys::ConnectLobby) != config->additionalParameters.end() ? config->additionalParameters.at(ConfigurationKeys::ConnectLobby) : "";
 					_steamApiInitialize = config->additionalParameters.find(ConfigurationKeys::SteamApiInitialize) != config->additionalParameters.end() ? (config->additionalParameters.at(ConfigurationKeys::SteamApiInitialize) != "false") : true;
 					_steamApiRunCallbacks = config->additionalParameters.find(ConfigurationKeys::SteamApiRunCallbacks) != config->additionalParameters.end() ? (config->additionalParameters.at(ConfigurationKeys::SteamApiRunCallbacks) != "false") : true;
 
-					if (_connectLobby.empty() && config->processLaunchArguments.size() > 0)
+					if (_connectLobby.empty() && config->processLaunchArguments.size() >= 2)
 					{
 						for (auto argi = 0; argi < config->processLaunchArguments.size(); argi++)
 						{
-							if (config->processLaunchArguments[argi] == "+connect_lobby" && (argi + 1) < config->processLaunchArguments.size())
+							if (config->processLaunchArguments[argi] == "+connect_lobby" && config->processLaunchArguments.size() > (argi + 1))
 							{
 								std::string steamIDLobby = config->processLaunchArguments[argi + 1];
+
+								logger->log(LogLevel::Trace, "Steam", "Extract and store connectLobby from processLaunchArguments", steamIDLobby);
+
 								_connectLobby = steamIDLobby;
 							}
 						}
@@ -333,9 +336,9 @@ namespace Stormancer
 			{
 			public:
 
-				SteamPartyInvitation(const std::string& senderId, const Party::PartyId& partyId)
-					: _senderId(senderId)
-					, _partyId(partyId)
+				SteamPartyInvitation(const Party::PartyId& partyId, const std::string& senderSteamID = "")
+					: _partyId(partyId)
+					, _senderSteamID(senderSteamID)
 				{
 				}
 
@@ -351,7 +354,7 @@ namespace Stormancer
 
 				std::string getSenderId() override
 				{
-					return _senderId;
+					return _senderSteamID;
 				}
 
 				std::string getSenderPlatformId() override
@@ -366,8 +369,8 @@ namespace Stormancer
 
 			private:
 
-				std::string _senderId;
 				Party::PartyId _partyId;
+				std::string _senderSteamID;
 			};
 
 			class SteamPartyProvider;
@@ -413,6 +416,26 @@ namespace Stormancer
 						if (steamConfig->getSteamApiRunCallbacks())
 						{
 							scheduleRunSteamAPiCallbacks();
+						}
+
+						auto connectLobbyArgument = steamConfig->getConnectLobby();
+
+						if (!connectLobbyArgument.empty())
+						{
+							if (auto invitationMessenger = _wInvitationMessenger.lock())
+							{
+								_logger->log(LogLevel::Trace, "Steam", "Process launch argument '+connect_lobby'", connectLobbyArgument);
+
+								SteamIDLobby steamIDLobby = std::stoull(connectLobbyArgument);
+
+								Party::PartyId partyId;
+								partyId.id = std::to_string(steamIDLobby);
+								partyId.type = PARTY_TYPE_STEAMIDLOBBY;
+								partyId.platform = platformName;
+
+								auto steamPartyInvitation = std::make_shared<SteamPartyInvitation>(partyId);
+								invitationMessenger->notifyInvitationReceived(steamPartyInvitation);
+							}
 						}
 					}
 
@@ -1311,9 +1334,9 @@ namespace Stormancer
 				SteamID senderId = callback->m_steamIDFriend.ConvertToUint64();
 
 				Party::PartyId partyId;
-				partyId.platform = platformName;
-				partyId.type = PARTY_TYPE_STEAMIDLOBBY;
 				partyId.id = std::to_string(steamIDLobby);
+				partyId.type = PARTY_TYPE_STEAMIDLOBBY;
+				partyId.platform = platformName;
 
 				auto invitationMessenger = _wInvitationMessenger.lock();
 				if (!invitationMessenger)
@@ -1322,7 +1345,7 @@ namespace Stormancer
 					return;
 				}
 
-				auto steamPartyInvitation = std::make_shared<SteamPartyInvitation>(std::to_string(senderId), partyId);
+				auto steamPartyInvitation = std::make_shared<SteamPartyInvitation>(partyId, std::to_string(senderId));
 				invitationMessenger->notifyInvitationReceived(steamPartyInvitation);
 			}
 
@@ -1409,6 +1432,8 @@ namespace Stormancer
 						STORM_RETURN_TASK_FROM_EXCEPTION(std::runtime_error("Unknown PartyId type"), Party::PartyId);
 					}
 
+					_logger->log(LogLevel::Trace, "SteamPartyProvider::getPartyId", "Retrieve partyId from Steam lobby metadata", partyId.id);
+
 					std::lock_guard<std::recursive_mutex> lg(_mutex);
 
 					auto steamApi = _wSteamApi.lock();
@@ -1448,9 +1473,12 @@ namespace Stormancer
 								throw std::runtime_error("Invalid partyId");
 							}
 
+							logger->log(LogLevel::Trace, "SteamPartyProvider::getPartyId", "PartyId obtained", partyDataDto.partyId);
+
 							Party::PartyId partyId;
 							partyId.id = partyDataDto.partyId;
 							partyId.type = Party::PartyId::TYPE_PARTY_ID;
+
 							return partyId;
 						});
 					});
@@ -1933,7 +1961,7 @@ namespace Stormancer
 
 			void registerClientDependencies(ContainerBuilder& builder) override
 			{
-				builder.registerDependency<details::SteamState, Configuration>().singleInstance();
+				builder.registerDependency<details::SteamState, Configuration, ILogger>().singleInstance();
 				builder.registerDependency<details::SteamImpl, Users::UsersApi, details::SteamState, Configuration, IScheduler, ILogger, Party::PartyApi, Party::Platform::InvitationMessenger>().asSelf().as<SteamApi>().singleInstance();
 				builder.registerDependency<details::SteamPartyProvider, Party::Platform::InvitationMessenger, Users::UsersApi, details::SteamImpl, ILogger, Party::PartyApi, IActionDispatcher>().as<Party::Platform::IPlatformSupportProvider>();
 				builder.registerDependency<SteamAuthenticationEventHandler, details::SteamState>().as<Users::IAuthenticationEventHandler>();